% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_d}
\alias{compute_d}
\title{this function calculates Sm (1 by p), sum of z in event set; O, the vector of total events in each
facility; num_events, number of all events; events_per_day_facility, number of events each day in each facility}
\usage{
compute_d(hosp_begin, max_d, D, z, facility, num_facility)
}
\arguments{
\item{hosp_begin}{matrix of hospitalization begin date}

\item{max_d}{the largest number of hospitalization in each segment}

\item{D}{number of distinct failure time number of days)}

\item{z}{time-independent variables}
}
\description{
this function calculates Sm (1 by p), sum of z in event set; O, the vector of total events in each
facility; num_events, number of all events; events_per_day_facility, number of events each day in each facility
}
