% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_d_omp}
\alias{compute_d_omp}
\title{this function calculates Sm (1 by p), sum of z in event set; events_per_facility, the vector of total events in each
facility; num_events, number of all events; events_per_day_facility, number of events each day in each facility}
\usage{
compute_d_omp(
  hosp_begin,
  max_d,
  D,
  z,
  Z_tv,
  t_tv,
  max_v,
  facility,
  num_facility,
  nthreads = 1L
)
}
\arguments{
\item{hosp_begin}{matrix of hospitalization begin date}

\item{max_d}{the largest number of hospitalization in each segment}

\item{D}{number of distinct failure time number of days)}

\item{z}{time-independent variables}
}
\description{
this function calculates Sm (1 by p), sum of z in event set; events_per_facility, the vector of total events in each
facility; num_events, number of all events; events_per_day_facility, number of events each day in each facility
}
