% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cox_breslow_dep}
\alias{cox_breslow_dep}
\title{Estimate beta, lambda, and alpha}
\usage{
cox_breslow_dep(
  hosp_begin,
  max_d,
  t_start,
  t_end,
  z,
  Z_tv,
  t_tv,
  max_v,
  tolb,
  tola,
  maxiter,
  facility,
  num_facility,
  D,
  beta,
  parallel = TRUE,
  nthreads = 1L
)
}
\arguments{
\item{max_d}{the largest number of hospitalization in each segment}

\item{z}{design matrix}

\item{facility}{1 to num_facility}

\item{delta}{indicator vector of event}

\item{tol}{tolerance of precision, when highest change in beta is lower than or equal to tol, iteration stops}
}
\value{
estimated beta, lambda, alpha
}
\description{
Estimate beta, lambda, and alpha
}
