% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cox_breslow_ind}
\alias{cox_breslow_ind}
\title{Estimate beta, lambda, and alpha}
\usage{
cox_breslow_ind(
  hosp_begin,
  max_d,
  t_start,
  t_end,
  z,
  tolb,
  tola,
  maxiter,
  facility,
  num_facility,
  D,
  beta
)
}
\arguments{
\item{hosp_begin}{matrix of event time for each participant}

\item{max_d}{number of events for each row (participant)}

\item{z}{design matrix}

\item{facility}{1 to num_facility}

\item{beta}{initial value for beta}

\item{delta}{indicator vector of event}

\item{tol}{tolerance of precision, when highest change in beta is lower than or equal to tol, iteration stops}
}
\value{
estimated beta, lambda, alpha
}
\description{
Estimate beta, lambda, and alpha
}
