% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{fe.data.prep}
\alias{fe.data.prep}
\title{Check Whether the Input is Valid}
\usage{
fe.data.prep(data, T1.char, T2.char, event.prefix = NULL,
  event.num = NULL, event.char = NULL, nevents.char = NULL, Zti.char,
  Ztd.char = NULL, ttd.char = NULL, prov.char, TD = TRUE,
  threshold.cor = 0.9, cutoff = 10, check = FALSE)
}
\arguments{
\item{data}{a data frame including response, provider ID, and covariates, with missing values imputed}

\item{T1.char}{a character string as name of time to enter the study}

\item{T2.char}{a character string as name of time to exit the study}

\item{event.prefix}{a character string as the prefix of name of event time variable}

\item{event.num}{maximum number of events for one patient}

\item{event.char}{a character string as the prefix of name of event time variable}

\item{nevents.char}{a character string as name of number of events of each patient}

\item{Zti.char}{a vector of character strings as names of time-independent covariates}

\item{Ztd.char}{a vector of character strings as names of time-dependent covariates}

\item{ttd.char}{a vector of character strings as names of change points of time-dependent covariates}

\item{prov.char}{a character string as name of variable consisting of provider IDs}

\item{TD}{a Boolean indicating whether there are time-dependent covariates, default true}

\item{threshold.cor}{minimum pairwise correlation among time-independent covariates to trigger a warning}

\item{cutoff}{an integer as cutoff of provider size with 10 as default. A provider will be filtered out if its
size is smaller or equal to cutoff}

\item{check}{a Boolean (default FALSE) indicating whether checks of variation, correlation, collinearity are
needed}

\item{id.char}{a character string as name of patient ID}
}
\value{
A data frame with non-missing patients and non-small facilities only and additional variables 'no.readm', which
equals 1 if the block has no events. Notice that the block effects of those with no events are truncated at -10.
}
\description{
Check absence of variables, missingness of variables, variation in covariates, correlation, and VIF.
If we fit a linear model of explanatory variable j vs. all other explanatory variables,
and calculate the model's R-square, then the VIF of variable j is an increasing function of
this R-square: $VIF_j=1/(1-R_j^2)$. In other words, VIF evaluates how much one explanatory variable can be
 explained by other explanatory variables in a linear relationship. The more a variable is linearly
 related to other variables, the higher the VIF of this variable. Usually, a VIF larger than 10 is
 considered as having high collinearity.
}
