% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_indep.R
\name{generate_data}
\alias{generate_data}
\title{Generate Data with Time-independent Covariates}
\usage{
generate_data(
  true_beta = c(c(0.5, -0.5, 1, -1, 1.5), rep(0, 5)),
  F_pre = 1:10,
  gamma = rep(1, 10),
  dmu0 = 0.002,
  ndays = 200,
  seed = 23021619,
  mean_sample_size = 50,
  Sigma_z1 = 0.1 * diag(length(true_beta)),
  time_depen = TRUE
)
}
\arguments{
\item{true_beta}{true parameters}

\item{F_pre}{a unique list of facilities}

\item{gamma}{true facility effects of each facility, notice length of gamma must be the same as length of F_pre}

\item{dmu0}{a constant for baseline rate}

\item{ndays}{number of days followed}

\item{seed}{random seed}

\item{mean_sample_size}{expected sample size of each facility (stratum), a scalar same in all facilities}

\item{Sigma_z1}{Var-Cov matrix of design matrix}

\item{time_depen}{indicator of whether the dataset contains time-dependent covariates. This function generates 2 time-dependent covariates for now}
}
\value{
a list of generated data:
\item{z}{design matrix}
\item{facility}{which facility each patient belongs to}
\item{true_beta}{true parameters}
\item{t_start}{enter time}
\item{t_end}{exit time}
\item{num_facility}{number of facilities (strata)}
\item{events_time}{observed event time for each patient}
\item{nevents}{number of events of each patient}
\item{ndays}{number of follow-up days}
\item{N}{number of patients}
}
\description{
The function generate_data_indep generates example recurrent events data with time-independent covariates only
}
\examples{
surv_dat = generate_data()

}
