% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{recur.fe.prov}
\alias{recur.fe.prov}
\title{Fit Recurrent Events Model}
\usage{
recur.fe.prov(
  data,
  T1.char,
  T2.char,
  event.char,
  id.char,
  nevents.char,
  Zti.char,
  Ztd.char = NULL,
  ttd.char = NULL,
  prov.char,
  tolb = 1e-06,
  tola = 1e-06,
  bound = 10,
  max.iter = 10000,
  backtrack = FALSE,
  TimeDepen = TRUE,
  parallel = FALSE,
  AUC = FALSE
)
}
\arguments{
\item{data}{a data frame including response, provider ID, and covariates, with missing values imputed}

\item{T1.char}{a character string as name of time to enter the study}

\item{T2.char}{a character string as name of time to exit the study}

\item{event.char}{a character string as the prefix of name of event time variable}

\item{id.char}{a character string as name of patient ID}

\item{nevents.char}{a character string as name of number of events of each patient}

\item{Zti.char}{a vector of character strings as names of time-independent covariates}

\item{Ztd.char}{a vector of character strings as names of time-dependent covariates}

\item{ttd.char}{a vector of character strings as names of change points of time-dependent covariates}

\item{prov.char}{a character string as name of variable consisting of provider IDs}

\item{tolb}{a small positive number specifying stopping criterion of Newton-Raphson algorithm for beta: covariate effect}

\item{tola}{a small positive number specifying stopping criterion of Newton-Raphson algorithm for alpha: block effect}

\item{bound}{the maximum number that block effect can take}

\item{max.iter}{maximum number of iterations of beta fitting}

\item{backtrack}{a boolean indicating whether backtracking line search is implemented, defaulting to FALSE}

\item{TimeDepen}{a Boolean indicating whether there are time-dependent covariates, default true}

\item{parallel}{a Boolean indicating whether the parallel computing by OpenMP is used, default false}
}
\value{
a data frame with non-missing patients and non-small facilities only and additional variables 'no.readm', which
equals 1 if the block has no events. Notice that the block effects of those with no events are truncated at -10
}
\description{
Fit a recurrent events model with different options
}
